refactor(credit): modularize kyc_pipeline_screening → kyc/ package (Wave 1 PR #3)
Converts 3 flat files (kyc_pipeline_screening.py 621 LOC + kyc_models.py 41 LOC
+ kyc_client.py 94 LOC) into kyc/ package with 7 files:

  kyc/
    models.py           — KYCScreening, KYCScreeningMatch stubs (leaf)
    client.py           — KYCSpiderClient (httpx, no siblings)
    entity_extraction.py — person + org extraction (cross-package: sponsor/)
    screening.py        — individual person/org screening
    appendix.py         — Markdown appendix builder
    persistence.py      — DB persistence (imports models.py)
    service.py          — run_kyc_screenings orchestrator (never-raises)
    __init__.py         — standard imports, __all__

Never-raises contract: service.py wraps all exceptions.
stdlib logging → structlog.
Cross-package dep: entity_extraction imports sponsor.extract_key_persons_from_analysis.
test_vertical_engines.py: 3 entries → 1 package entry.
