+++ docs/plans/2026-03-15-refactor-credit-vertical-modular-alignment-wave1-plan.md
@@ -178,7 +178,7 @@ sponsor/
 
 ---
 
-### Phase 3: kyc/ (PR #3)
+### Phase 3: kyc/ (PR #3) ✅
 
 **Source:** `kyc_pipeline_screening.py` (621 LOC) + `kyc_models.py` (30 LOC) + `kyc_client.py` (70 LOC)
